nullb/benchmarks: add strace output
basic sync r/w test to compare the performance
of native vs isolated driver.

this is not the real implementation to be benchmarked because
there is only one ipc inside lcd during data exchange, request
completions are handled from klcd layer. This gives a rough idea
on where we are heading to. Looks like there is more work to do!

Signed-off-by: Abhiram Balasubramanian <abhiram@cs.utah.edu>