block-mq: Introduce async handling of requests
Additional hook to the block mq layer allows to
pass the list of requests in a single shot to the
klcd so that we can perform async communication with
the lcd. If otherwise, we have take the complete async
logic into the kernel (libasync) so that DO_FINISH and
ASYNC macros are visible inside the kernel. This can be
done if libasync is moved into the kernel/lib/.

Signed-off-by: Abhiram Balasubramanian <abhiram@cs.utah.edu>