kliblcd: Add lcd_exit which takes task_struct
This is needed for cleaning up resources like cptr_cache,
itree, lcd that was created for a thread which came in
from user-space. This function is useful when async threads
come in from the block layer which doesn't invoke a block
device open. So we store the info when a thread lands
in klcd module inside a list and we parse the list to
cleanup all the resources during device close.

Signed-off-by: Abhiram Balasubramanian <abhiram@cs.utah.edu>