kliblcd: Add helpers to create character device
chardev interface is needed for the user-space
to share a region through mmap system call. This
shared region will be used for user application like
fio to allocate data so that when direct-IO operations
are performed, the memory is already available inside
the lcd. As direct-IO incurs no overhead of copying
in the kernel, this makes sense for accessing data pointers
that arrive inside bios without any copies inside the LCD.

Signed-off-by: Abhiram Balasubramanian <abhiram@cs.utah.edu>